Update and fixes from Ryan Lakanen
- moved the npc logging into the function used when interacting with an
NPC. It will log all NPCs you can talk to, including vendors and quest
givers. But, of course, not the NPCs you can't talk to. Better than
nothing, though.

- Commented out the event listener for watching the reticleover unit
and also commented out the remainder of provisioning code that was
still functional.

- commented out the reticleover code, just in case anything changes in
the future so it will still be there.